backend/requirements.txt
fastapi
uvicorn
pdfplumber
requests
cohere
python-dotenv
supabase
